+++ src/main/java/com/mgen256/al/blocks/FirePitBase.java
@@ -1,7 +1,7 @@
 package com.mgen256.al.blocks;
 
-import net.minecraft.block.Block;
-import net.minecraft.util.math.shapes.VoxelShape;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.phys.shapes.VoxelShape;
 
 
 public abstract class FirePitBase extends Pedestal  {