+++ src/main/java/com/mgen256/al/items/SoulWand.java
@@ -3,27 +3,23 @@ package com.mgen256.al.items;
 import com.mgen256.al.FireTypes;
 import com.mgen256.al.blocks.*;
 
-import net.minecraft.block.Block;
-import net.minecraft.block.BlockState;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.item.Item;
-import net.minecraft.item.ItemStack;
-import net.minecraft.util.ActionResult;
-import net.minecraft.util.ActionResultType;
-import net.minecraft.util.Hand;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.BlockRayTraceResult;
-import net.minecraft.util.math.RayTraceResult;
-import net.minecraft.util.math.RayTraceContext.FluidMode;
-import net.minecraft.world.World;
+import net.minecraft.core.BlockPos;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.InteractionResultHolder;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.ClipContext;
 
 public class SoulWand extends Wand {
     
     private static Properties createProps(){
         Properties p = new Item.Properties();
         p.setNoRepair();
-        p.maxStackSize(1);
-        p.defaultMaxDamage(1);
+        p.stacksTo(1);
+        p.defaultDurability(1);
         return p;
     }
 
@@ -34,58 +30,54 @@ public class SoulWand extends Wand {
     }
 
     @Override
-    public ActionResult<ItemStack> onItemRightClick(World worldIn, PlayerEntity playerIn, Hand handIn)
+    public InteractionResultHolder<ItemStack> use(Level level, Player player, InteractionHand handIn)
     {
-        ItemStack stack = playerIn.getHeldItem(handIn);
-
-        RayTraceResult rayTraceResult = rayTrace(worldIn, playerIn, FluidMode.NONE);
-        BlockRayTraceResult blockRayTraceResult = (BlockRayTraceResult) rayTraceResult;
-
-        BlockPos pos = blockRayTraceResult.getPos();
-        BlockState state = worldIn.getBlockState(pos);
-        Block block = state.getBlock();
+        var stack = player.getItemInHand(handIn);
+        var hitresult = getPlayerPOVHitResult(level, player, ClipContext.Fluid.NONE);
+        var pos = hitresult.getBlockPos();
+        var state = level.getBlockState(pos);
+        var block = state.getBlock();
 
         if( block instanceof IHasFire )
         {
-            changeFire( worldIn, playerIn, pos, state, (IHasFire)block );
+            changeFire( level, player, pos, state, (IHasFire)block );
         }
         else if( block instanceof FireBase )
         {
-            BlockPos underPos = pos.down();
-            Block underBlock = worldIn.getBlockState( underPos ).getBlock();
+            var underPos = pos.below();
+            var underBlock = level.getBlockState( underPos ).getBlock();
             if( underBlock instanceof IHasFire )
-                changeFire( worldIn, playerIn, underPos, worldIn.getBlockState(underPos), (IHasFire)underBlock );
+                changeFire( level, player, underPos, level.getBlockState(underPos), (IHasFire)underBlock );
         }
-      
-        return new ActionResult<ItemStack>(ActionResultType.CONSUME, stack);
+        return InteractionResultHolder.consume(stack);
     }
 
 
-    private void changeFire( World worldIn, PlayerEntity playerIn, BlockPos pos, BlockState state, IHasFire modblock ) {
+    private void changeFire( Level level, Player player, BlockPos pos, BlockState state, IHasFire modblock ) {
 
-        FireTypes currentType = state.get( IHasFire.FIRE_TYPE );
-        FireTypes prevType = state.get( IHasFire.PREVIOUS_FIRE_TYPE );
+        FireTypes currentType = state.getValue( IHasFire.FIRE_TYPE );
+        FireTypes prevType = state.getValue( IHasFire.PREVIOUS_FIRE_TYPE );
         if( prevType == FireTypes.SOUL )
             prevType = FireTypes.NORMAL;
 
-        if( playerIn.isSneaking() )
+        if( player.isSuppressingSlidingDownLadder() )
         {
             if( currentType == FireTypes.SOUL )
             {
-                state = modblock.setFireType( worldIn, pos, state, prevType, prevType );
-                playSound( worldIn, playerIn, SoundEvents.UNDO, 0.6f );
+                state = modblock.setFireType( level, pos, state, prevType, prevType );
+                playSound( level, player, SoundEvents.UNDO, 0.6f );
             }
         }
         else
         {
             if( currentType != FireTypes.SOUL )
             {
-                state = modblock.setFireType( worldIn, pos, state, FireTypes.SOUL, prevType );
-                playSound( worldIn, playerIn, SoundEvents.CHANGE, 0.8f );      
+                state = modblock.setFireType( level, pos, state, FireTypes.SOUL, prevType );
+                playSound( level, player, SoundEvents.CHANGE, 0.8f );      
             }
         }
 
         if( modblock instanceof Pedestal )
-            ((Pedestal)modblock).setFire(worldIn, pos, state, true );
+            ((Pedestal)modblock).setFire(level, pos, state, true );
     }
 }