+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
 
-## [1.16.4-2.1.3] - 2020-11-18
-- Update to 1.16.4
+## [1.17.1-2.1.4] - 2021-08-30
+- Update to 1.17.1
+- Changed the recipe for Fire Pit L and S (deleted one block column and changed the coal to two).
+- Fixed an issue where fire blocks were not broken when breaking blocks that were immersed in water.
+- Fixed the dark display of Wand in GUI.
+- Compatible with forge 1.17.1-37.0.31